gbsddialog: implement --save-theme
This is meant to be the opposite to --load-theme, in that it should
output the CSS with Gtk+ 3, or the RC file with Gtk+ 2.

Overall this is not compatible with the original behaviour from
bsddialog though, and it may make sense to re-implement this to be
closer to the original behaviour.